+++ modulo5/projeto-todolist/.gitignore
@@ -1,0 +1,4 @@
+node_modules
+package-lock.json
+build
+.env
