--- modulo5/projeto-todolist/knex-express-template/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "aula-knex",
-  "version": "1.0.0",
-  "main": "index.js",
-  "scripts": {
-    "start": "tsc && node ./build/index.js",
-    "dev-start": "ts-node-dev ./src/index.ts"
-  },
-  "author": "Labenu",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "knex": "^0.95.2",
-    "mysql": "^2.18.1",
-    "uuid": "^9.0.0"
-  },
-  "devDependencies": {
-    "@types/cors": "^2.8.10",
-    "@types/express": "^4.17.11",
-    "@types/knex": "^0.16.1",
-    "@types/node": "^14.14.35",
-    "@types/uuid": "^8.3.4",
-    "ts-node-dev": "^1.1.6",
-    "typescript": "^4.2.3"
-  }
-}
