--- modulo5/projeto-todolist/knex-express-template/tsconfig.json
@@ -1,14 +1,0 @@
-{
-   "compilerOptions": {
-      "target": "es6",
-      "module": "commonjs",
-      "sourceMap": true,
-      "outDir": "./build",
-      "rootDir": "./src",
-      "removeComments": true,
-      "strict": true,
-      "noImplicitAny": true,
-      "esModuleInterop": true,
-      "forceConsistentCasingInFileNames": true
-   }
-}
