+++ modulo5/projeto-todolist/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+      "target": "es2021",       
+      "module": "commonjs",   
+      "sourceMap": true,       
+      "outDir": "./build",      
+      "rootDir": "./src",       
+      "removeComments": true,   
+      "noImplicitAny": true,      
+      "esModuleInterop": true
+  }
+}
